% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color}
\alias{color}
\alias{color_pattern}
\title{Functions to extract IMPACT colors as hex codes}
\usage{
color(..., unname = TRUE)

color_pattern(pattern, unname = TRUE)
}
\arguments{
\item{...}{Character names of colors. If NULL returns all colors.}

\item{unname}{Boolean. Should the output vector be unnamed? Default to \code{TRUE}.}

\item{pattern}{Pattern of the start of colors' name.}
}
\value{
Hex codes named or unnamed.
}
\description{
\code{\link[=color]{color()}} returns the requested columns, returns NA if absent. \code{\link[=color_pattern]{color_pattern()}} returns all colors that start with the pattern.
}
\section{Naming of colors}{

\itemize{
\item All branding colors start with "branding";
\item All categorical colors start with "cat_";
\item All continuous colors start with "quant_";
}

Then, a number indicates the number of colors that belong to the palettes, a string the name of the palette, and, finally, a number the position of the color. E.g., "quant_5_red_4" would be the 4th color of a continuous palettes of 5 colors in the red band. Exception is made for white, light_grey, dark_grey, and black.
}

