% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactR.viz-package.R
\docType{package}
\name{impactR.viz-package}
\alias{impactR.viz}
\alias{impactR.viz-package}
\title{impactR.viz: Color and viz helpers, wrappers, templates}
\description{
It basically provides colors as hex codes, color palettes, and some viz functions (graphs and maps).
}
\author{
\strong{Maintainer}: Noblet Guillaume \email{gnoblet@zaclys.net}

}
\keyword{internal}
