% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_gen.R
\name{palette_gen}
\alias{palette_gen}
\alias{palette_gen_discrete}
\alias{palette_gen_continuous}
\title{Generate color palettes}
\usage{
palette_gen(palette, type, direction = 1, ...)

palette_gen_discrete(palette = "branding_reach", direction = 1)

palette_gen_continuous(palette = "quant_7_artichoke", direction = 1, ...)
}
\arguments{
\item{palette}{Palette name from \code{\link[=palette]{palette()}}.}

\item{type}{"discrete" or "continuous".}

\item{direction}{1 or -1; should the order of colors be reversed?}

\item{...}{Additional arguments to pass to \code{\link[=colorRampPalette]{colorRampPalette()}} when type is "continuous".}
}
\description{
\code{\link[=palette_gen]{palette_gen()}} generates a color palette and let you choose whether continuous or discrete. \code{\link[=palette_gen_discrete]{palette_gen_discrete()}} and \code{\link[=palette_gen_continuous]{palette_gen_continuous()}} generates respectively discrete and continuous palettes.
}
