% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_color_impact_discrete}
\alias{scale_color_impact_discrete}
\alias{scale_fill_impact_discrete}
\alias{scale_fill_impact_continuous}
\alias{scale_color_impact_continuous}
\title{Scale constructors for fill and colors}
\usage{
scale_color_impact_discrete(
  palette = "cat_5_ibm",
  direction = 1,
  reverse_guide = TRUE,
  ...
)

scale_fill_impact_discrete(
  palette = "cat_5_ibm",
  direction = 1,
  reverse_guide = TRUE,
  ...
)

scale_fill_impact_continuous(
  palette = "quant_7_artichoke",
  direction = 1,
  reverse_guide = TRUE,
  ...
)

scale_color_impact_continuous(
  palette = "quant_7_artichoke",
  direction = 1,
  reverse_guide = TRUE,
  ...
)
}
\arguments{
\item{palette}{Palette name from \code{\link[=palette]{palette()}}.}

\item{direction}{1 or -1; should the order of colors be reversed?}

\item{reverse_guide}{Boolean indicating whether the guide should be reversed.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} if discrete or \code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradient()}} if continuous.}
}
\description{
This function is based on \code{\link[=palette]{palette()}}. If palette is NULL, the used palette will be magma from gpplot2's viridis scale constructors.
}
